MVA-119
feat: refactor LicenciaDeConducirComponent to use sonner for toast notifications

chore: remove hidden route for editing generic services in sidebar data

chore: add @radix-ui/react-toast dependency for toast notifications

feat: add employee leave types constants

feat: implement utility functions for date formatting and service/leave status badges

feat: create custom hook for employee data management

feat: create custom hook for leave management with form handling

feat: create custom hook for logout functionality

feat: create custom hook for service management with task handling

feat: define types for dashboard employee services and leave management
