+++ components/sections/constants/employeeConstants.ts
@@ -1,0 +1,9 @@
+export const availableLeaveTypes = [
+  { value: "VACACIONES", label: "Vacaciones" },
+  { value: "ENFERMEDAD", label: "Licencia por enfermedad" },
+  { value: "CASAMIENTO", label: "Licencia por casamiento" },
+  { value: "NACIMIENTO", label: "Licencia por nacimiento" },
+  { value: "FALLECIMIENTO_FAMILIAR", label: "Licencia por fallecimiento" },
+  { value: "CAPACITACION", label: "Licencia por capacitación" },
+  { value: "ORDINARIA", label: "Licencia ordinaria" },
+];
